Add parsing of tables with bare expressions
The syntax described in the manual is the following:

    tableconstructor ::= ‘{’ [fieldlist] ‘}’
    fieldlist ::= field {fieldsep field} [fieldsep]

Unfortunately, describing it like that with Menhir (LR(1)) produced a
shift/reduce conflict.  This is caused by the (optional) field separator
at the end.  That's why I used a slightly different grammar to describe
the same syntax.

Also, this made some AST types mutually recursive.  I'm not sure I'll
keep them in dedicated sub modules since mutually recursive modules are
quite verbose and complex.